minecraftutils/requirements.txt
Flask==0.10.1
Flask-Classy==0.6.10
Flask-WTF==0.9.5
PyYAML==3.11